[Roles] SuperUser forged

+++ Genesis/Modules/Roles/Roles.ascx
@@ -37,20 +37,19 @@
                     {
                         foreach (Models.Mod_Roles Role in LMR)
                         {
-                            Response.Write(
+                            string buttons = "";
+                            
+                            if(Role.ROLE != "SuperUser")
+                            {
+                                buttons = "<input style='opacity: 0;' type='button' class='ButtonDark' value='X' onclick='DeleteRole(" + Role.ID + ")' />" +
+                                "<input style='opacity: 0;' class='ButtonDark' type='button' value='EDITAR' onclick='GetRoleToEdit(this," + Role.ID + ")'/>";
+
+                                Response.Write(
                                 "<tr  style='cursor: grab;' class='draggable RowRole TableRows' idRole=" + Role.ID + ">" +
-                                    "<td>" + Role.ROLE +
-                                    //"<label class='InputTextDarkLabel'>" +
-                                    //    "<input idRole='" + Role.ID + "' style='color: rgba(255,255,255,0.8)' placeholder='ej: Enfermerx' class='InputTextDark RoleName' type='text' value='" + Role.ROLE + "'/>" +
-                                    //    "<span class='InputTextDarkPlaceholderWrap'> <span class='InputTextDarkPlaceholder'>" + Role.ROLE + "</span> </span>" +
-                                    //"</label>" +
-                                        
-                                    "</td>" +
-                                    "<td>" + 
-                                        "<input style='opacity: 0;' type='button' class='ButtonDark' value='X' onclick='DeleteRole(" + Role.ID + ")' />" +
-                                        "<input style='opacity: 0;' class='ButtonDark' type='button' value='EDITAR' onclick='GetRoleToEdit(this," + Role.ID + ")'/>" +
-                                    "</td>" +
+                                    "<td>" + Role.ROLE + "</td>" +
+                                    "<td>" + buttons + "</td>" +
                                 "</tr>");
+                            }                            
                         }
                     }
                     else
@@ -81,7 +80,8 @@
 
                                     foreach (Models.Mod_Roles Roles in User.ROLES)
                                     {
-                                        listItems += "<li> <input type='button' class='ButtonDark' value='X' onclick='UnassignRole(\"" + Roles.ID + "\",\"" + User.ID + "\")' /> " + Roles.ROLE + " </li>";
+                                        if(Roles.ROLE != "SuperUser")
+                                            listItems += "<li> <input type='button' class='ButtonDark' value='X' onclick='UnassignRole(\"" + Roles.ID + "\",\"" + User.ID + "\")' /> " + Roles.ROLE + " </li>";
                                     }
                                 }
